% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilExtractor.R
\name{def_loc}
\alias{def_loc}
\title{Extract soil property data from SoilGrids.}
\usage{
def_loc(x, y, z)
}
\arguments{
\item{x}{Longitude in decimal degrees}

\item{y}{Latitude in decimal degrees}

\item{z}{}
}
\value{
spatialpointsdataframe
}
\description{
Create a spatial point dataframe.
}
\examples{
x <- 37.437971
y <- -1.86698
z <-"Matwiku"
pnts <- def_loc(x,y, z)

}
