% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilExtractor.R
\name{defineLocation}
\alias{defineLocation}
\title{Create a spatial point dataframe}
\usage{
defineLocation(x, y, z)
}
\arguments{
\item{x}{Longitude in decimal degrees}

\item{y}{Latitude in decimal degrees}

\item{z}{Character containing name of location, ID etc}
}
\description{
Create a spatial point dataframe
}
\examples{
library(grime)
x <- 37.437971
y <- -1.86698
z <-"Matwiku"
pnts <- defineLocation(x,y, z)
pnts

}
