% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilExtractor.R
\name{defineSoilProperties}
\alias{defineSoilProperties}
\title{Create class for SoilGrids REST API}
\usage{
defineSoilProperties(...)
}
\arguments{
\item{ls}{Vector containing list of soil properties to extract (ORCDRC, PHIHOX)}
}
\value{
list
}
\description{
Create class for SoilGrids REST API
}
\examples{
library(grime)
ls <- c("ORCDRC", "PHIHOX")
slist <- defineSoilProperties(ls)
slist

}
