% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilExtractor.R
\name{extractSoilsProperties}
\alias{extractSoilsProperties}
\title{Extract soil properties from soil grids}
\usage{
extractSoilsProperties(x, y)
}
\arguments{
\item{slist}{Object of class "list"; contains parameters or REST.SoilGrids query}

\item{pnts}{Geometry (locations) of the queries}
}
\value{
dataframe
}
\description{
Extract soil properties from soil grids
}
\examples{
library(grime)
x <- 37.437971
y <- -1.86698
z <-"Matwiku"
pnts <- defineLocation(x,y,z)
a <- c("ORCDRC", "PHIHOX")
slist <- defineSoilProperties(a)
soil_properties <- extractSoilsProperties(slist, pnts)
soil_properties

}
