% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilExtractor.R
\name{defineLocation}
\alias{defineLocation}
\title{Create a spatial point dataframe}
\usage{
defineLocation(lon, lat, id)
}
\arguments{
\item{lon}{Longitude in decimal degrees}

\item{lat}{Latitude in decimal degrees}

\item{id}{Character containing name of location, ID etc}
}
\description{
Create a spatial point dataframe
}
\examples{
library(sp)
lon <- 37.437971
lat <- -1.86698
id <-"Matwiku"
pnts <- defineLocation(lon, lat, id)
pnts

}
