% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilExtractor.R
\name{extractSoilsProperties}
\alias{extractSoilsProperties}
\title{Extract soil properties from soil grids}
\usage{
extractSoilsProperties(slist, pnts)
}
\arguments{
\item{slist}{Object of class "list"; contains parameters or REST.SoilGrids query}

\item{pnts}{Geometry (locations) of the queries}
}
\value{
dataframe
}
\description{
Extract soil properties from soil grids
}
\examples{
library(grime)
lon <- 37.437971
lat <- -1.86698
id <-"Matwiku"
pnts <- defineLocation(lon, lat, id)
ls <- c("ORCDRC", "PHIHOX")
slist <- defineSoilProperties(ls)
soil_properties <- extractSoilsProperties(slist, pnts)
soil_properties

}
